Fix mod meters in PA/RT

+++ pahandler.h
@@ -80,7 +80,7 @@ private:
     quint32         lastSentSeq = 0;
 
     quint16 currentLatency;
-    quint16 amplitude = 0;
+    float amplitude=0.0;
     qreal volume = 1.0;
 
     audioSetup setup;
